refactor(routes): extract jellyfin action routes
